Moved BarrettGlaser action specific code to new location.
This required a change to how we 'install' the header files, so that the directory structure is preserved.
Refactored the Doxygen cmake code to tidy the main CMakeLists.txt file.
