+++ RFL_source/new_source/tests/tAction.cpp
@@ -1,7 +1,7 @@
 //
 // Created by Paul Druce on 27/12/2022.
 //
-#include "Action.hpp"
+#include "BarrettGlaser/Action.hpp"
 #include "GslRng.hpp"
 #include <gtest/gtest.h>
 
@@ -84,13 +84,3 @@ TEST(ActionTests, CreateWithNoParams) {
       EXPECT_EQ(action.getG2(), 0);
       EXPECT_EQ(action.getG4(), 0););
 }
-
-//TEST(ActionTests, PrintAction) {
-//  DiracOperator dirac(1, 1, 5);
-//  Action action(2.0, 4.0);
-//
-//  std::stringstream captured_stream;
-//  action.printS(dirac, captured_stream);
-//
-//  EXPECT_EQ(captured_stream.str(), "200 800\n");
-//}
