+++ RFL_source/source/CMakeLists.txt
@@ -20,7 +20,8 @@ set_target_properties(RFL PROPERTIES PUBLIC_HEADER "${RFL_INCLUDES}")
 target_include_directories(RFL PUBLIC ${RFL_INCLUDE_DIR})
 target_link_libraries(RFL ${ARMADILLO_LIBRARIES} ${GSL_LIBRARIES})
 
+# Install static library and header files
 install(TARGETS RFL
-        DESTINATION ${CMAKE_SOURCE_DIR}/lib/RFL/bin
+        DESTINATION ${CMAKE_SOURCE_DIR}/lib/RFL/bin/${CMAKE_BUILD_TYPE}
         PUBLIC_HEADER DESTINATION ${CMAKE_SOURCE_DIR}/lib/RFL/include
         )
